introduced RestClient and moved rest template interactions to there. Similar chage on the test cases too
