+++ data-prepper-plugins/saas-source-plugins/jira-source/src/test/java/org/opensearch/dataprepper/plugins/source/jira/JiraServiceTest.java
@@ -6,30 +6,19 @@ import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
-import org.junit.jupiter.params.provider.MethodSource;
-import org.junit.jupiter.params.provider.ValueSource;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.opensearch.dataprepper.plugins.source.jira.exception.BadRequestException;
-import org.opensearch.dataprepper.plugins.source.jira.exception.UnAuthorizedException;
 import org.opensearch.dataprepper.plugins.source.jira.models.IssueBean;
 import org.opensearch.dataprepper.plugins.source.jira.models.SearchResults;
-import org.opensearch.dataprepper.plugins.source.jira.rest.auth.JiraAuthConfig;
-import org.opensearch.dataprepper.plugins.source.jira.rest.auth.JiraAuthFactory;
+import org.opensearch.dataprepper.plugins.source.jira.rest.JiraRestClient;
 import org.opensearch.dataprepper.plugins.source.source_crawler.base.PluginExecutorServiceProvider;
 import org.opensearch.dataprepper.plugins.source.source_crawler.model.ItemInfo;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.client.HttpClientErrorException;
-import org.springframework.web.client.RestTemplate;
 
 import java.io.IOException;
 import java.io.InputStream;
-import java.net.URI;
 import java.time.Instant;
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -38,7 +27,6 @@ import java.util.Map;
 import java.util.Queue;
 import java.util.UUID;
 import java.util.concurrent.ConcurrentLinkedQueue;
-import java.util.stream.Stream;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
@@ -52,14 +40,11 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.when;
 import static org.opensearch.dataprepper.plugins.source.jira.utils.Constants.ACCESSIBLE_RESOURCES;
-import static org.opensearch.dataprepper.plugins.source.jira.utils.Constants.AUTHORIZATION_ERROR_CODE;
 import static org.opensearch.dataprepper.plugins.source.jira.utils.Constants.BASIC;
 import static org.opensearch.dataprepper.plugins.source.jira.utils.Constants.CREATED;
 import static org.opensearch.dataprepper.plugins.source.jira.utils.Constants.KEY;
 import static org.opensearch.dataprepper.plugins.source.jira.utils.Constants.NAME;
-import static org.opensearch.dataprepper.plugins.source.jira.utils.Constants.OAUTH2;
 import static org.opensearch.dataprepper.plugins.source.jira.utils.Constants.PROJECT;
-import static org.opensearch.dataprepper.plugins.source.jira.utils.Constants.TOKEN_EXPIRED;
 import static org.opensearch.dataprepper.plugins.source.jira.utils.Constants.UPDATED;
 
 
@@ -71,13 +56,10 @@ public class JiraServiceTest {
 
     private static final Logger log = LoggerFactory.getLogger(JiraServiceTest.class);
     private final PluginExecutorServiceProvider executorServiceProvider = new PluginExecutorServiceProvider();
-    @Mock
-    private RestTemplate restTemplate;
-    @Mock
-    private JiraAuthConfig authConfig;
 
     @Mock
-    private StringBuilder jql;
+    private JiraRestClient jiraRestClient;
+
 
     private static InputStream getResourceAsStream(String resourceName) {
         InputStream inputStream = Thread.currentThread().getContextClassLoader().getResourceAsStream(resourceName);
@@ -97,13 +79,24 @@ public class JiraServiceTest {
         return null;
     }
 
-    private static Stream<Arguments> provideHttpStatusCodesWithExceptionClass() {
-        return Stream.of(
-                Arguments.of(HttpStatus.valueOf(AUTHORIZATION_ERROR_CODE), UnAuthorizedException.class),
-                Arguments.of(HttpStatus.valueOf(TOKEN_EXPIRED), RuntimeException.class),
-                Arguments.of(HttpStatus.TOO_MANY_REQUESTS, RuntimeException.class),
-                Arguments.of(HttpStatus.INSUFFICIENT_STORAGE, RuntimeException.class)
-        );
+    public static JiraSourceConfig createJiraConfiguration(String auth_type,
+                                                           List<String> issueType,
+                                                           List<String> issueStatus,
+                                                           List<String> projectKey) throws JsonProcessingException {
+        ObjectMapper objectMapper = new ObjectMapper();
+        Map<String, String> connectorCredentialsMap = new HashMap<>();
+        connectorCredentialsMap.put("auth_type", auth_type);
+
+        Map<String, Object> jiraSourceConfigMap = new HashMap<>();
+        jiraSourceConfigMap.put("account_url", ACCESSIBLE_RESOURCES);
+        jiraSourceConfigMap.put("connector_credentials", connectorCredentialsMap);
+        jiraSourceConfigMap.put("issue_types", issueType);
+        jiraSourceConfigMap.put("statuses", issueStatus);
+        jiraSourceConfigMap.put("projects", projectKey);
+
+
+        String jiraSourceConfigJsonString = objectMapper.writeValueAsString(jiraSourceConfigMap);
+        return objectMapper.readValue(jiraSourceConfigJsonString, JiraSourceConfig.class);
     }
 
     @AfterEach
@@ -117,7 +110,7 @@ public class JiraServiceTest {
         List<String> issueStatus = new ArrayList<>();
         List<String> projectKey = new ArrayList<>();
         JiraSourceConfig jiraSourceConfig = createJiraConfiguration(BASIC, issueType, issueStatus, projectKey);
-        JiraService jiraService = new JiraService(restTemplate, jiraSourceConfig, authConfig);
+        JiraService jiraService = new JiraService(jiraSourceConfig, jiraRestClient);
         assertNotNull(jiraService);
     }
 
@@ -130,7 +123,7 @@ public class JiraServiceTest {
         issueStatus.add("Done");
         projectKey.add("KAN");
         JiraSourceConfig jiraSourceConfig = createJiraConfiguration(BASIC, issueType, issueStatus, projectKey);
-        JiraService jiraService = spy(new JiraService(restTemplate, jiraSourceConfig, authConfig));
+        JiraService jiraService = spy(new JiraService(jiraSourceConfig, jiraRestClient));
         List<IssueBean> mockIssues = new ArrayList<>();
         IssueBean issue1 = createIssueBean(false);
         mockIssues.add(issue1);
@@ -141,7 +134,7 @@ public class JiraServiceTest {
         when(mockSearchResults.getIssues()).thenReturn(mockIssues);
         when(mockSearchResults.getTotal()).thenReturn(mockIssues.size());
 
-        doReturn(mockSearchResults).when(jiraService).getAllIssues(any(StringBuilder.class), anyInt(), any(JiraSourceConfig.class));
+        doReturn(mockSearchResults).when(jiraRestClient).getAllIssues(any(StringBuilder.class), anyInt(), any(JiraSourceConfig.class));
 
         Instant timestamp = Instant.ofEpochSecond(0);
         Queue<ItemInfo> itemInfoQueue = new ConcurrentLinkedQueue<>();
@@ -157,7 +150,7 @@ public class JiraServiceTest {
         List<String> projectKey = new ArrayList<>();
         issueType.add("Task");
         JiraSourceConfig jiraSourceConfig = createJiraConfiguration(BASIC, issueType, issueStatus, projectKey);
-        JiraService jiraService = spy(new JiraService(restTemplate, jiraSourceConfig, authConfig));
+        JiraService jiraService = spy(new JiraService(jiraSourceConfig, jiraRestClient));
         List<IssueBean> mockIssues = new ArrayList<>();
         for (int i = 0; i < 50; i++) {
             IssueBean issue1 = createIssueBean(false);
@@ -168,7 +161,7 @@ public class JiraServiceTest {
         when(mockSearchResults.getIssues()).thenReturn(mockIssues);
         when(mockSearchResults.getTotal()).thenReturn(100);
 
-        doReturn(mockSearchResults).when(jiraService).getAllIssues(any(StringBuilder.class), anyInt(), any(JiraSourceConfig.class));
+        doReturn(mockSearchResults).when(jiraRestClient).getAllIssues(any(StringBuilder.class), anyInt(), any(JiraSourceConfig.class));
 
         Instant timestamp = Instant.ofEpochSecond(0);
         Queue<ItemInfo> itemInfoQueue = new ConcurrentLinkedQueue<>();
@@ -188,7 +181,7 @@ public class JiraServiceTest {
         projectKey.add("!@#$");
 
         JiraSourceConfig jiraSourceConfig = createJiraConfiguration(BASIC, issueType, issueStatus, projectKey);
-        JiraService jiraService = new JiraService(restTemplate, jiraSourceConfig, authConfig);
+        JiraService jiraService = new JiraService(jiraSourceConfig, jiraRestClient);
 
         Instant timestamp = Instant.ofEpochSecond(0);
         Queue<ItemInfo> itemInfoQueue = new ConcurrentLinkedQueue<>();
@@ -202,98 +195,13 @@ public class JiraServiceTest {
         List<String> projectKey = new ArrayList<>();
         issueType.add("Task");
         JiraSourceConfig jiraSourceConfig = createJiraConfiguration(BASIC, issueType, issueStatus, projectKey);
-        JiraService jiraService = spy(new JiraService(restTemplate, jiraSourceConfig, authConfig));
+        JiraService jiraService = spy(new JiraService(jiraSourceConfig, jiraRestClient));
 
-        doThrow(RuntimeException.class).when(jiraService).getAllIssues(any(StringBuilder.class), anyInt(), any(JiraSourceConfig.class));
+        doThrow(RuntimeException.class).when(jiraRestClient).getAllIssues(any(StringBuilder.class), anyInt(), any(JiraSourceConfig.class));
 
         Instant timestamp = Instant.ofEpochSecond(0);
         Queue<ItemInfo> itemInfoQueue = new ConcurrentLinkedQueue<>();
         assertThrows(RuntimeException.class, () -> jiraService.getJiraEntities(jiraSourceConfig, timestamp, itemInfoQueue));
-    }
-
-    @Test
-    public void testGetAllIssuesBasic() throws JsonProcessingException {
-        List<String> issueType = new ArrayList<>();
-        List<String> issueStatus = new ArrayList<>();
-        List<String> projectKey = new ArrayList<>();
-        issueType.add("Task");
-        JiraSourceConfig jiraSourceConfig = createJiraConfiguration(BASIC, issueType, issueStatus, projectKey);
-        JiraService jiraService = new JiraService(restTemplate, jiraSourceConfig, authConfig);
-        SearchResults mockSearchResults = mock(SearchResults.class);
-        doReturn(new ResponseEntity<>(mockSearchResults, HttpStatus.OK)).when(restTemplate).getForEntity(any(URI.class), any(Class.class));
-        SearchResults results = jiraService.getAllIssues(jql, 0, jiraSourceConfig);
-        assertNotNull(results);
-    }
-
-    @Test
-    public void testGetAllIssuesOauth2() throws JsonProcessingException {
-        List<String> issueType = new ArrayList<>();
-        List<String> issueStatus = new ArrayList<>();
-        List<String> projectKey = new ArrayList<>();
-        issueType.add("Task");
-        JiraSourceConfig jiraSourceConfig = createJiraConfiguration(OAUTH2, issueType, issueStatus, projectKey);
-        JiraService jiraService = new JiraService(restTemplate, jiraSourceConfig, authConfig);
-        SearchResults mockSearchResults = mock(SearchResults.class);
-        doReturn("http://mock-service.jira.com").when(authConfig).getUrl();
-        doReturn(new ResponseEntity<>(mockSearchResults, HttpStatus.OK)).when(restTemplate).getForEntity(any(URI.class), any(Class.class));
-        SearchResults results = jiraService.getAllIssues(jql, 0, jiraSourceConfig);
-        assertNotNull(results);
-    }
-
-    private JiraSourceConfig createJiraConfiguration(String auth_type, List<String> issueType, List<String> issueStatus, List<String> projectKey) throws JsonProcessingException {
-        ObjectMapper objectMapper = new ObjectMapper();
-        Map<String, String> connectorCredentialsMap = new HashMap<>();
-        connectorCredentialsMap.put("auth_type", auth_type);
-
-        Map<String, Object> jiraSourceConfigMap = new HashMap<>();
-        jiraSourceConfigMap.put("account_url", ACCESSIBLE_RESOURCES);
-        jiraSourceConfigMap.put("connector_credentials", connectorCredentialsMap);
-        jiraSourceConfigMap.put("issue_types", issueType);
-        jiraSourceConfigMap.put("statuses", issueStatus);
-        jiraSourceConfigMap.put("projects", projectKey);
-
-
-        String jiraSourceConfigJsonString = objectMapper.writeValueAsString(jiraSourceConfigMap);
-        return objectMapper.readValue(jiraSourceConfigJsonString, JiraSourceConfig.class);
-    }
-
-    @Test
-    void testInvokeRestApiTokenExpiredInterruptException() throws JsonProcessingException, InterruptedException {
-        List<String> issueType = new ArrayList<>();
-        List<String> issueStatus = new ArrayList<>();
-        List<String> projectKey = new ArrayList<>();
-        JiraSourceConfig jiraSourceConfig = createJiraConfiguration(BASIC, issueType, issueStatus, projectKey);
-        JiraService jiraService = new JiraService(restTemplate, jiraSourceConfig, authConfig);
-        when(authConfig.getUrl()).thenReturn("https://example.com/rest/api/2/issue/key");
-        when(restTemplate.getForEntity(any(URI.class), any(Class.class))).thenThrow(new HttpClientErrorException(HttpStatus.TOO_MANY_REQUESTS));
-        jiraService.setSleepTimeMultiplier(100000);
-
-        Thread testThread = new Thread(() -> {
-            assertThrows(InterruptedException.class, () -> {
-                try {
-                    jiraService.getIssue("key");
-                } catch (Exception e) {
-                    throw new RuntimeException(e);
-                }
-            });
-        });
-        testThread.start();
-        Thread.sleep(100);
-        testThread.interrupt();
-    }
-
-    @ParameterizedTest
-    @MethodSource("provideHttpStatusCodesWithExceptionClass")
-    void testInvokeRestApiTokenExpired(HttpStatus statusCode, Class expectedExceptionType) throws JsonProcessingException {
-        List<String> issueType = new ArrayList<>();
-        List<String> issueStatus = new ArrayList<>();
-        List<String> projectKey = new ArrayList<>();
-        JiraSourceConfig jiraSourceConfig = createJiraConfiguration(BASIC, issueType, issueStatus, projectKey);
-        JiraService jiraService = new JiraService(restTemplate, jiraSourceConfig, authConfig);
-        jiraService.setSleepTimeMultiplier(1);
-        when(authConfig.getUrl()).thenReturn("https://example.com/rest/api/2/issue/key");
-        when(restTemplate.getForEntity(any(URI.class), any(Class.class))).thenThrow(new HttpClientErrorException(statusCode));
-        assertThrows(expectedExceptionType, () -> jiraService.getIssue("key"));
     }
 
     private IssueBean createIssueBean(boolean nullFields) {
@@ -336,18 +244,6 @@ public class JiraServiceTest {
         issue1.setFields(fieldMap);
 
         return issue1;
-    }
-
-    @ParameterizedTest
-    @ValueSource(strings = {"basic-auth-jira-pipeline.yaml"})
-    public void testFetchingJiraIssue(String configFileName) {
-        String exampleTicketResponse = "{\"id\":\"123\",\"key\":\"key\",\"self\":\"https://example.com/rest/api/2/issue/123\"}";
-        doReturn(new ResponseEntity<>(exampleTicketResponse, HttpStatus.OK)).when(restTemplate).getForEntity(any(URI.class), any(Class.class));
-        JiraSourceConfig jiraSourceConfig = createJiraConfigurationFromYaml(configFileName);
-        JiraAuthConfig authConfig = new JiraAuthFactory(jiraSourceConfig).getObject();
-        JiraService jiraService = new JiraService(restTemplate, jiraSourceConfig, authConfig);
-        String ticketDetails = jiraService.getIssue("key");
-        assertEquals(exampleTicketResponse, ticketDetails);
     }
 
 }
